refactor: align Entity model boundaries
