+++ src/entities/card/api/id.ts
@@ -1,5 +1,4 @@
-import type { CardId } from "./types";
-
 import { generateId } from "@/shared/lib/generateId";
+import type { CardId } from "../model/types";
 
 export const generateCardId = (): CardId => generateId();
