+++ src/entities/card/index.ts
@@ -1,7 +1,7 @@
 export { fetchCards, subscribeCards } from "./api/firestore";
+export { generateCardId } from "./api/id";
 export { CardBulkMutationError, deleteCard, editCard, mutateCards } from "./api/mutations";
 export type { CardMutation } from "./api/mutations";
-export { generateCardId } from "./model/id";
 export { useCard, useCards, useCardsByDeckId } from "./model/hooks";
 export { cardContentSchema } from "./model/schema";
 export { clearRemoteCards } from "./model/store";
