+++ src/entities/deck/api/id.ts
@@ -1,5 +1,4 @@
-import type { DeckId } from "./types";
-
 import { generateId } from "@/shared/lib/generateId";
+import type { DeckId } from "../model/types";
 
 export const generateDeckId = (): DeckId => generateId();
