+++ src/entities/study-session/model/store.ts
@@ -145,12 +145,8 @@ export const removeStudySession = (deckId: DeckId): void => {
   });
 };
 
-export const clearStudySessions = (): Promise<void> => {
-  // The executor runs immediately while converting synchronous cleanup failures into Promise rejections.
-  return new Promise((resolve) => {
-    // Publish the empty state before durable cleanup so auth changes cannot expose the previous user's sessions.
-    studySessionStore.setState({ sessionsByDeckId: {} });
-    studySessionStore.persist.clearStorage();
-    resolve();
-  });
+export const clearStudySessions = (): void => {
+  // Publish the empty state before durable cleanup so auth changes cannot expose the previous user's sessions.
+  studySessionStore.setState({ sessionsByDeckId: {} });
+  studySessionStore.persist.clearStorage();
 };
